user_service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
bcrypt==4.0.1
pymysql
cryptography